Fix clippy 1.93 warnings (CI -Dwarnings)
CI started failing under clippy 1.93 with:

  unnecessary_cast in frontend/src/main.rs (e.loaded() as f64 -- both
                  loaded() and total() already return f64)
  unused_imports  for `bail` in backend/src/decode.rs (left over after
                  switching from bail! to Context-based error returns)
  unnecessary_map_or in backend/src/decode.rs (map_or(false, _) ->
                  is_ok_and on the FITS HDU search closure)

All three are pure cleanups -- no behavior change.

+++ frontend/src/main.rs
@@ -52,7 +52,7 @@ async fn upload_with_progress(
     let upload = xhr.upload().map_err(|e| format!("{e:?}"))?;
     let progress_cb = Closure::wrap(Box::new(move |e: ProgressEvent| {
         if e.length_computable() {
-            let pct = e.loaded() as f64 / e.total() as f64;
+            let pct = e.loaded() / e.total();
             on_progress(pct);
         }
     }) as Box<dyn FnMut(ProgressEvent)>);
